web/requirements.txt
Django==1.11
Faker==0.7.16
bpython
celery==4.0.2
django-filter==1.0.4
djangorestframework==3.6.3
factory_boy==2.8.1
gunicorn==19.7.1
isort
mypy
psycopg2==2.7.1
selenium==2.53.6
soundcloud
